software/requirements.txt
bidict==0.23.1
bleak==0.22.3
bluez-peripheral==0.1.7
dbus-next==0.2.3
dbus-python==1.3.2
dnspython==2.7.0
eventlet==0.38.0
greenlet==3.1.1
h11==0.14.0
hostp2pd==0.1.7
lockfile==0.12.2
netifaces==0.11.0
python-daemon==3.1.0
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
setuptools==75.6.0
simple-websocket==1.1.0
typing_extensions==4.12.2
wsproto==1.2.0
